Move pip, poetry, conda, code to deployment package. Simplify pip freeze support.
